+++ src/Audit.NET.SqlServer/ConfigurationApi/ISqlServerProviderConfigurator.cs
@@ -3,6 +3,9 @@ using Audit.Core;
 #if NETSTANDARD1_3 || NETSTANDARD2_0 || NETSTANDARD2_1 || NET5_0
 using Microsoft.EntityFrameworkCore;
 #endif
+#if NET45
+using System.Data.Common;
+#endif
 
 namespace Audit.SqlServer.Configuration
 {
@@ -17,6 +20,11 @@ namespace Audit.SqlServer.Configuration
         /// To set the database initializer to NULL on the internal DbContext 
         /// </summary>
         ISqlServerProviderConfigurator SetDatabaseInitializerNull(bool initializeToNull = true);
+        /// <summary>
+        /// Specifies the Sql Server DbConnection to use when creating the DB Context. 
+        /// Use this instead of `ConnectionString()` when you require low level configuration of the connection.
+        /// </summary>
+        ISqlServerProviderConfigurator DbConnection(Func<AuditEvent, DbConnection> connection);
 #else
         /// <summary>
         /// Specifies the DB context options as a function of the audit event.