+++ src/Audit.NET.SqlServer/ConfigurationApi/SqlServerProviderConfigurator.cs
@@ -6,6 +6,9 @@ using Audit.Core;
 #if NETSTANDARD1_3 || NETSTANDARD2_0 || NETSTANDARD2_1 || NET5_0
 using Microsoft.EntityFrameworkCore;
 #endif
+#if NET45
+using System.Data.Common;
+#endif
 
 namespace Audit.SqlServer.Configuration
 {
@@ -29,6 +32,14 @@ namespace Audit.SqlServer.Configuration
             _setDatabaseInitializerNull = initializeToNull;
             return this;
         }
+
+        internal Func<AuditEvent, DbConnection> _dbConnectionBuilder { get; set; }
+        public ISqlServerProviderConfigurator DbConnection(Func<AuditEvent, DbConnection> connection)
+        {
+            _dbConnectionBuilder = connection;
+            return this;
+        }
+
 #endif
 
         public ISqlServerProviderConfigurator ConnectionString(string connectionString)