+++ test/Audit.EntityFramework.Core.UnitTest/BlogsContext.cs
@@ -28,7 +28,9 @@ namespace Audit.EntityFramework.Core.UnitTest
             optionsBuilder.UseLoggerFactory(_loggerFactory);
             optionsBuilder.UseSqlServer(CnnString);
             optionsBuilder.EnableSensitiveDataLogging();
+#if !NETCOREAPP1_0
             optionsBuilder.UseLazyLoadingProxies();
+#endif
         }
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {