+++ test/Audit.EntityFramework.Core.UnitTest/DepartmentContext.cs
@@ -21,7 +21,9 @@ namespace Audit.EntityFramework.Full.UnitTest
         {
             optionsBuilder.UseSqlServer(CnnString);
             optionsBuilder.EnableSensitiveDataLogging();
+#if !NETCOREAPP1_0
             optionsBuilder.UseLazyLoadingProxies();
+#endif
         }
 
         public override int SaveChanges()
@@ -49,7 +51,9 @@ namespace Audit.EntityFramework.Full.UnitTest
         {
             optionsBuilder.UseSqlServer(CnnString);
             optionsBuilder.EnableSensitiveDataLogging();
+#if !NETCOREAPP1_0
             optionsBuilder.UseLazyLoadingProxies();
+#endif
         }
 
         public override int SaveChanges()